# backend/shortVideoApiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /user/code:
        post:
            tags:
                - UserService
            description: 获取验证码
            operationId: UserService_GetVerificationCode
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.GetVerificationCodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.GetVerificationCodeResponse'
    /user/info:
        get:
            tags:
                - UserService
            description: 获取用户信息
            operationId: UserService_GetUserInfo
            parameters:
                - name: userId
                  in: query
                  description: '@gotags: json:"user_id,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.GetUserInfoResponse'
        put:
            tags:
                - UserService
            description: 更新用户信息
            operationId: UserService_UpdateUserInfo
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.UpdateUserInfoRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.UpdateUserInfoResponse'
    /user/login:
        post:
            tags:
                - UserService
            description: 登录
            operationId: UserService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.LoginResponse'
    /user/register:
        post:
            tags:
                - UserService
            description: 注册
            operationId: UserService_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.RegisterResponse'
    /user/voucher:
        post:
            tags:
                - UserService
            operationId: UserService_BindUserVoucher
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.BindUserVoucherRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.BindUserVoucherResponse'
        delete:
            tags:
                - UserService
            operationId: UserService_UnbindUserVoucher
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/svapi.UnbindUserVoucherRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/svapi.UnbindUserVoucherResponse'
...
components:
    schemas:
...
        svapi.BindUserVoucherRequest:
            type: object
            properties:
                voucherType:
                    type: integer
                    format: enum
                voucher:
                    type: string
        svapi.BindUserVoucherResponse:
            type: object
            properties: {}
...
        svapi.GetUserInfoResponse:
            type: object
            properties:
                user:
                    $ref: '#/components/schemas/svapi.User'
        svapi.GetVerificationCodeRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
        svapi.GetVerificationCodeResponse:
            type: object
            properties:
                codeId:
                    type: string
                    description: '@gotags: json:"code_id,omitempty,string"'
...
        svapi.LoginRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
                password:
                    type: string
        svapi.LoginResponse:
            type: object
            properties:
                token:
                    type: string
...
        svapi.RegisterRequest:
            type: object
            properties:
                mobile:
                    type: string
                email:
                    type: string
                password:
                    type: string
                codeId:
                    type: string
                    description: '@gotags: json:"code_id,omitempty,string"'
                code:
                    type: string
        svapi.RegisterResponse:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
...
        svapi.UnbindUserVoucherRequest:
            type: object
            properties:
                voucherType:
                    type: integer
                    format: enum
                voucher:
                    type: string
        svapi.UnbindUserVoucherResponse:
            type: object
            properties: {}
...
        svapi.UpdateUserInfoRequest:
            type: object
            properties:
                userId:
                    type: string
                    description: '@gotags: json:"user_id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
        svapi.UpdateUserInfoResponse:
            type: object
            properties: {}
        svapi.User:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                backgroundImage:
                    type: string
                signature:
                    type: string
                mobile:
                    type: string
                email:
                    type: string
                followCount:
                    type: string
                    description: '@gotags: json:"followCount,omitempty,string"'
                followerCount:
                    type: string
                    description: '@gotags: json:"followerCount,omitempty,string"'
                totalFavorited:
                    type: string
                    description: '@gotags: json:"totalFavorited,omitempty,string"'
                workCount:
                    type: string
                    description: '@gotags: json:"workCount,omitempty,string"'
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favoriteCount,omitempty,string"'
...